Progress on quotient knowls
