+++ lmfdb/groups/abstract/templates/abstract-show-group.html
@@ -339,7 +339,7 @@
   <tr>
     <td>{{KNOWL('group.center',title='Center')}}:</td>
     <td><span class="subgp chargp" data-sgid="{{gp.cent()}}">$Z \simeq {{gp.cent_label()}}$</span></td>
-    <td>$G/Z \simeq {{gp.central_quot()}}$</td>
+    <td>$G/Z \simeq$ {{gp.central_quot()|safe}}</td>
   </tr>
   <tr>
     <td>{{KNOWL('group.commutator_subgroup',title='Commutator')}}:</td>
@@ -350,22 +350,22 @@
     <tr>
     <td>{{KNOWL('group.frattini_subgroup',title='Frattini')}}:</td>
     <td><span class="subgp chargp" data-sgid="{{gp.fratt()}}">$\Phi \simeq {{gp.fratt_label()}}$</span></td>
-    <td>$G/\Phi \simeq {{gp.frattini_quot()}}$ </td>
+    <td>$G/\Phi \simeq$ {{gp.frattini_quot()|safe}} </td>
   </tr>
  <tr>
     <td>{{KNOWL('group.fitting_subgroup',title='Fitting')}}:</td>
     <td><span class="subgp chargp" data-sgid="{{gp.fitting}}">$\operatorname{Fit} \simeq {{gp.subgroups[gp.fitting].subgroup_tex}}$</td>
-    <td>$G/\operatorname{Fit} \simeq {{gp.subgroups[gp.fitting].quotient_tex}}$</td>
+    <td>$G/\operatorname{Fit} \simeq$ {{info.display_group_knowl(gp.subgroups[gp.fitting].quotient)|safe}}</td>
   </tr>
   <tr>
     <td>{{KNOWL('group.radical',title='Radical')}}:</td>
     <td><span class="subgp chargp" data-sgid="{{gp.radical}}">$R \simeq {{gp.subgroups[gp.radical].subgroup_tex}}$</td>
-    <td>$G/R \simeq {{gp.subgroups[gp.radical].quotient_tex}}$</td>
+    <td>$G/R \simeq$ {{info.display_group_knowl(gp.subgroups[gp.radical].quotient)|safe}}</td>
   </tr>
   <tr>
     <td>{{KNOWL('group.socle',title='Socle')}}:</td>
     <td><span class="subgp chargp" data-sgid="{{gp.socle}}">$\operatorname{soc} \simeq {{gp.subgroups[gp.socle].subgroup_tex}}$</td>
-    <td>$G/\operatorname{soc} \simeq {{gp.subgroups[gp.socle].quotient_tex}}$</td>
+    <td>$G/\operatorname{soc} \simeq$ {{info.display_group_knowl(gp.subgroups[gp.socle].quotient)|safe}}</td>
   </tr>
   {% if gp.pgroup == 0 %}
     {% for p, subgp in gp.sylow_subgroups() %}
